add chat handling
