+++ src/parser/types.ts
@@ -21,6 +21,7 @@ import type {
 } from "./phase/phase-4";
 import type { Phase5Blocked, Phase5BlockedPlatforms, Phase5SocialInteract } from "./phase/phase-5";
 import type { Phase6RemoteItem, Phase6TxtEntry } from "./phase/phase-6";
+import type { Phase7Chat } from "./phase/phase-7";
 import {
   PhasePendingPodcastId,
   PhasePendingSocial,
@@ -180,6 +181,9 @@ export interface FeedObject extends BasicFeed {
   podcastRemoteItems?: Phase6RemoteItem[];
   podroll?: Phase6RemoteItem[];
   // #endregion
+  //
+  chat?: Phase7Chat;
+
   // #region Pending Phase
   /** PENDING AND LIKELY TO CHANGE indicates a listing on multiple platforms, directories, hosts, apps and services. */
   podcastId?: PhasePendingPodcastId[];
@@ -250,6 +254,9 @@ export interface Episode {
   podcastTxt?: Phase6TxtEntry[];
   podcastRemoteItems?: Phase6RemoteItem[];
   // #endregion
+
+  chat?: Phase7Chat;
+
   // #region Pending Phase
   podcastImages?: Phase4PodcastImage[];
   podcastRecommendations?: PhasePendingPodcastRecommendation[];
